fix: resolve build errors
